technical-document/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic
python-multipart==0.0.9
gitpython==3.1.43
tree-sitter
tree-sitter-language-pack
tiktoken
sentence-transformers
chromadb
python-dotenv
groq
python-docx
reportlab
atlassian-python-api
langchain
langgraph
numpy>=2.0